requirements.txt
--extra-index-url https://download.pytorch.org/whl/nightly/cu128
--pre
gymnasium==1.1.1
numpy==2.3.1
pandas==2.3.1
pandas_ta==0.3.14b0
PyYAML==6.0.2
rich==14.0.0
sb3_contrib @ git+https://github.com/akbaig/stable-baselines3-contrib@e722d75a417fc9612916e7446dc8e968f620801d
stable_baselines3==2.6.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tqdm==4.67.1
hydra-core~=1.3.2
omegaconf~=2.3.0
stable_baselines3~=2.6.0
sb3_contrib~=2.6.0a2